requirements.txt
# Core server
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.0.0
websockets>=12.0

# Audio processing
librosa>=0.10.0
numpy>=1.24.0

# Deep learning — CPU-only build (lighter, no CUDA deps, works on all servers)
tensorflow-cpu>=2.13.0

# YouTube / audio download
yt-dlp>=2024.1.0
imageio-ffmpeg>=0.4.9

# Music search / autocomplete
ytmusicapi>=1.5.0

# Training pipeline
scikit-learn>=1.3.0    # train/val split
